+++ src/workflows/registration.ts
@@ -1,24 +1,20 @@
 import { Signer } from '@ethersproject/abstract-signer';
-import {
-  serializeSignature,
-  sign,
-  signRaw,
-} from '../utils';
+import { serializeSignature, sign, signRaw } from '../utils';
 import {
   GetSignableRegistrationResponse,
   RegisterUserResponse,
   UsersApi,
 } from '../api';
 import { Registration } from '../contracts';
-import { StarkWallet } from '../types';
+import { L2Signer, StarkWallet } from '../types';
 
+/** @deprecated */
 export async function registerOffchainWorkflow(
   signer: Signer,
   starkWallet: StarkWallet,
   usersApi: UsersApi,
 ): Promise<RegisterUserResponse> {
-
-  const userAddress = (await signer.getAddress());
+  const userAddress = await signer.getAddress();
 
   // Get signable details for offchain registration
   const signableResult = await usersApi.getSignableRegistrationOffchain({
@@ -39,13 +35,53 @@ export async function registerOffchainWorkflow(
     sign(starkWallet.starkKeyPair, payloadHash),
   );
 
-  // Send request for user registratin offchain
+  // Send request for user registration offchain
   const response = await usersApi.registerUser({
     registerUserRequest: {
       eth_signature: ethSignature,
       ether_key: userAddress,
       stark_signature: starkSignature,
       stark_key: starkWallet.starkPublicKey,
+    },
+  });
+
+  return {
+    tx_hash: response.data.tx_hash,
+  };
+}
+
+export async function registerOffchainWorkflowWithSigner(
+  l1Signer: Signer,
+  l2Signer: L2Signer,
+  usersApi: UsersApi,
+): Promise<RegisterUserResponse> {
+  const userAddress = await l1Signer.getAddress();
+  const starkPublicKey = await l2Signer.getAddress();
+
+  // Get signable details for offchain registration
+  const signableResult = await usersApi.getSignableRegistrationOffchain({
+    getSignableRegistrationRequest: {
+      ether_key: userAddress,
+      stark_key: starkPublicKey,
+    },
+  });
+
+  const { signable_message: signableMessage, payload_hash: payloadHash } =
+    signableResult.data;
+
+  // Sign message with L1 credentials
+  const ethSignature = await signRaw(signableMessage, l1Signer);
+
+  // Sign hash with L2 credentials
+  const starkSignature = await l2Signer.signMessage(payloadHash);
+
+  // Send request for user registration offchain
+  const response = await usersApi.registerUser({
+    registerUserRequest: {
+      eth_signature: ethSignature,
+      ether_key: userAddress,
+      stark_signature: starkSignature,
+      stark_key: starkPublicKey,
     },
   });