+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Added `registerOffchainWithSigner` function to enable register offchain user with L2Signer
+
+### Changed
+- mark `registerOffchain` as deprecated
+
+### Fixed
+- `getAddress` method from BaseSigner
+
+## [0.4.0] - 2022-07-05
+### Added
 - Added `BaseSigner`, a default implementation of the Stark L2Signer interface
 - Added `createTradeWithSigner` function to enable creating trade workflow with l2signer